# feat: improve TS SDK runtime performances