--- a/sdk/typescript/package.json
+++ b/sdk/typescript/package.json
@@ -27,7 +27,7 @@
     "adm-zip": "^0.5.10",
     "env-paths": "^3.0.0",
     "execa": "^8.0.1",
-    "graphql": "^16.8.0",
+    "graphql": "^16.8.1",
     "graphql-request": "^6.1.0",
     "graphql-tag": "^2.12.6",
     "node-color-log": "^12.0.0",