--- a/sdk/typescript/package.json
+++ b/sdk/typescript/package.json
@@ -64,5 +64,6 @@
     "prettier": "^3.2.5",
     "ts-node": "^10.9.1",
     "typescript-eslint": "^7.8.0"
-  }
+  },
+  "packageManager": "yarn@4.2.2"
 }