--- a/sdk/typescript/runtime/main.go
+++ b/sdk/typescript/runtime/main.go
@@ -63,7 +63,7 @@ const (
 func (t *TypescriptSdk) ModuleRuntime(ctx context.Context, modSource *ModuleSource, introspectionJSON string) (*Container, error) {
 	ctr, err := t.CodegenBase(ctx, modSource, introspectionJSON)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("failed to create codegen base: %w", err)
 	}
 
 	subPath, err := modSource.SourceSubpath(ctx)
@@ -71,7 +71,7 @@ func (t *TypescriptSdk) ModuleRuntime(ctx context.Context, modSource *ModuleSour
 		return nil, fmt.Errorf("could not load module config: %w", err)
 	}
 
-	detectedRuntime, err := t.DetectRuntime(ctx, modSource, subPath)
+	detectedRuntime, err := t.detectRuntime(ctx, modSource, subPath)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create module runtime: %w", err)
 	}
@@ -106,7 +106,7 @@ func (t *TypescriptSdk) Codegen(ctx context.Context, modSource *ModuleSource, in
 	// Get base container
 	ctr, err := t.CodegenBase(ctx, modSource, introspectionJSON)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("failed to create codegen base: %w", err)
 	}
 	return dag.GeneratedCode(ctr.Directory(ModSourceDirPath)).
 		WithVCSGeneratedPaths([]string{
@@ -132,98 +132,32 @@ func (t *TypescriptSdk) CodegenBase(ctx context.Context, modSource *ModuleSource
 		return nil, fmt.Errorf("could not load module config: %w", err)
 	}
 
-	detectedRuntime, err := t.DetectRuntime(ctx, modSource, subPath)
+	detectedRuntime, err := t.detectRuntime(ctx, modSource, subPath)
 	if err != nil {
-		return nil, fmt.Errorf("failed to create codegen base: %w", err)
+		return nil, fmt.Errorf("failed to detect runtime: %w", err)
 	}
 
 	base, err := t.Base(detectedRuntime)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("failed to create codegen base: %w", err)
 	}
 
-	gen := base.
-		// Add codegen binary into a special path
-		WithMountedFile(codegenBinPath, t.SDKSourceDir.File("/codegen")).
-		// Add introspection file
-		WithNewFile(schemaPath, ContainerWithNewFileOpts{
-			Contents: introspectionJSON,
-		}).
-		// Execute the code generator using the given introspection file
-		WithExec([]string{
-			codegenBinPath,
-			"--lang", "typescript",
-			"--output", ModSourceDirPath,
-			"--module-name", name,
-			"--module-context-path", filepath.Join(ModSourceDirPath, subPath),
-			"--introspection-json-path", schemaPath,
-		}, ContainerWithExecOpts{
-			ExperimentalPrivilegedNesting: true,
-		}).
-		Directory(filepath.Join(ModSourceDirPath, subPath, GenDir))
+	// Get a directory with the SDK sources installed and the generated client.
+	sdk := t.
+		installedSDK(base, detectedRuntime).
+		WithDirectory(".", t.generateClient(base, name, introspectionJSON, subPath))
 
 	base = base.
-		// Add codegen binary into a special path
-		WithMountedFile(codegenBinPath, t.SDKSourceDir.File("/codegen")).
 		// Add template directory
 		WithMountedDirectory("/opt/module", dag.CurrentModule().Source().Directory(".")).
-		// Mount users' module
-		WithMountedDirectory(ModSourceDirPath, modSource.ContextDirectory()).
-		WithWorkdir(filepath.Join(ModSourceDirPath, subPath)).
-		// Add generated code
-		WithDirectory(GenDir, gen)
+		// Mount users' module with SDK sources and generated client in it.
+		WithMountedDirectory(ModSourceDirPath, modSource.
+			ContextDirectory().
+			WithDirectory(filepath.Join(subPath, GenDir), sdk),
+		).
+		WithWorkdir(filepath.Join(ModSourceDirPath, subPath))
 
-	moduleFiles, err := base.Directory(".").Entries(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("could not list rootfs entries: %w", err)
-	}
-	packageJSONExist := slices.Contains(moduleFiles, "package.json")
-
-	switch detectedRuntime {
-	case Bun:
-		// Check if the project has existing source:
-		// if it does: update the tsconfig in case dagger isn't part of the configuration
-		// if not: copy the template and replace QuickStart with the module name
-		if packageJSONExist {
-			base = base.WithExec([]string{"bun", "/opt/module/bin/__tsconfig.updator.ts"})
-		} else {
-			base = base.WithDirectory(".", base.Directory("/opt/module/template"), ContainerWithDirectoryOpts{Include: []string{"*.json"}})
-		}
-
-	case Node:
-		// Check if the project has existing source:
-		// if it does: update the tsconfig in case dagger isn't part of the configuration
-		// if not: copy the template and replace QuickStart with the module name
-		if packageJSONExist {
-			base = base.WithExec([]string{"tsx", "/opt/module/bin/__tsconfig.updator.ts"})
-		} else {
-			base = base.WithDirectory(".", base.Directory("/opt/module/template"), ContainerWithDirectoryOpts{Include: []string{"*.json"}})
-		}
-	default:
-		return nil, fmt.Errorf("unknown runtime: %s", detectedRuntime)
-	}
-
-	if !slices.Contains(moduleFiles, "src") {
-		base = base.WithDirectory("src", dag.Directory())
-	}
-
-	moduleSourceFiles, err := base.Directory("src").Entries(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("could not list module source entries: %w", err)
-	}
-
-	// Check if there's a src directory with .ts files in it.
-	// If not, add the template file and replace QuickStart with the module name
-	// This cover the case where there's a package.json but no src directory.
-	if !slices.ContainsFunc(moduleSourceFiles, func(s string) bool {
-		return path.Ext(s) == ".ts"
-	}) {
-		base = base.
-			WithDirectory("src", base.Directory("/opt/module/template/src"), ContainerWithDirectoryOpts{Include: []string{"*.ts"}}).
-			WithExec([]string{"sed", "-i", "-e", fmt.Sprintf("s/QuickStart/%s/g", strcase.ToCamel(name)), "src/index.ts"})
-	}
-
-	return base, nil
+	return t.setupModule(ctx, base, detectedRuntime, name)
 }
 
 // Base returns a Node or Bun container with cache setup for yarn or bun
@@ -233,15 +167,7 @@ func (t *TypescriptSdk) Base(runtime SupportedTSRuntime) (*Container, error) {
 		return dag.Container().
 			From(bunImageRef).
 			WithoutEntrypoint().
-			WithMountedCache("/root/.bun/install/cache", dag.CacheVolume(fmt.Sprintf("mod-bun-cache-%s", bunVersion))).
-			// Add SDK source code before to install its dependencies so we can benefit from the cache
-			// and do not reinstall the dependencies each time the codegen is run or something
-			// changed in the user file.
-			WithDirectory(GenDir, t.SDKSourceDir, ContainerWithDirectoryOpts{
-				Exclude: []string{"codegen", "runtime"},
-			}).
-			WithWorkdir(GenDir).
-			WithExec([]string{"bun", "install"}), nil
+			WithMountedCache("/root/.bun/install/cache", dag.CacheVolume(fmt.Sprintf("mod-bun-cache-%s", bunVersion))), nil
 	case Node:
 		return dag.Container().
 			From(nodeImageRef).
@@ -252,15 +178,108 @@ func (t *TypescriptSdk) Base(runtime SupportedTSRuntime) (*Container, error) {
 			WithEnvVariable("NODE_OPTIONS", "--use-openssl-ca").
 			WithMountedCache("/root/.npm", dag.CacheVolume(fmt.Sprintf("mod-npm-cache-%s", nodeVersion))).
 			WithMountedCache("/usr/local/share/.cache", dag.CacheVolume(fmt.Sprintf("mod-yarn-cache-%s", nodeVersion))).
-			WithExec([]string{"npm", "install", "-g", "tsx"}).
-			WithDirectory(GenDir, t.SDKSourceDir, ContainerWithDirectoryOpts{
-				Exclude: []string{"codegen", "runtime"},
-			}).
-			WithWorkdir(GenDir).
-			WithExec([]string{"yarn", "install", "--production"}), nil
+			WithExec([]string{"npm", "install", "-g", "tsx"}), nil
 	default:
 		return nil, fmt.Errorf("unknown runtime: %s", runtime)
 	}
+}
+
+// setupModule initialiaze the user's module.
+//
+// If the user's module has a package.json file, it will run the
+// __tsconfig.updator.ts script in order to add dagger to the tsconfig path so
+// the editor can give type hints and auto completion.
+// Otherwise, it will copy the template config files into the user's module directory.
+//
+// If there's no src directory or no typescript files in it, it will create one
+// and copy the template index.ts file in it.
+func (t *TypescriptSdk) setupModule(ctx context.Context, ctr *Container, runtime SupportedTSRuntime, name string) (*Container, error) {
+	moduleFiles, err := ctr.Directory(".").Entries(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("could not list rootfs entries: %w", err)
+	}
+
+	packageJSONExist := slices.Contains(moduleFiles, "package.json")
+
+	// If there's a package.json, run the tsconfig updator script.
+	// else, copy the template config files.
+	if packageJSONExist {
+		executor := "tsx"
+
+		if runtime == Bun {
+			executor = "bun"
+		}
+
+		ctr = ctr.WithExec([]string{executor, "/opt/module/bin/__tsconfig.updator.ts"})
+	} else {
+		ctr = ctr.WithDirectory(".", ctr.Directory("/opt/module/template"), ContainerWithDirectoryOpts{Include: []string{"*.json"}})
+	}
+
+	// Check if there's a src directory and creates an empty directory if it doesn't exist.
+	if !slices.Contains(moduleFiles, "src") {
+		ctr = ctr.WithDirectory("src", dag.Directory())
+	}
+
+	// Get the list of files in the src directory.
+	moduleSourceFiles, err := ctr.Directory("src").Entries(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("could not list module source entries: %w", err)
+	}
+
+	// Check if there's a src directory with .ts files in it.
+	// If not, add the template file and replace QuickStart with the module name
+	if !slices.ContainsFunc(moduleSourceFiles, func(s string) bool {
+		return path.Ext(s) == ".ts"
+	}) {
+		return ctr.
+			WithDirectory("src", ctr.Directory("/opt/module/template/src"), ContainerWithDirectoryOpts{Include: []string{"*.ts"}}).
+			WithExec([]string{"sed", "-i", "-e", fmt.Sprintf("s/QuickStart/%s/g", strcase.ToCamel(name)), "src/index.ts"}), nil
+	}
+
+	return ctr, nil
+}
+
+// installedSDK returns a directory with the SDK sources and its dependencies installed.
+func (t *TypescriptSdk) installedSDK(ctr *Container, runtime SupportedTSRuntime) *Directory {
+	ctr = ctr.
+		WithWorkdir(ModSourceDirPath).
+		WithDirectory(".", t.SDKSourceDir, ContainerWithDirectoryOpts{
+			Exclude: []string{"codegen", "runtime"},
+		})
+
+	switch runtime {
+	case Bun:
+		return ctr.WithExec([]string{"bun", "install"}).Directory(ModSourceDirPath)
+	case Node:
+		return ctr.WithExec([]string{"yarn", "install", "--production"}).Directory(ModSourceDirPath)
+	default:
+		// Should never happen since we verify the runtime before calling this function.
+		return nil
+	}
+}
+
+// generateClient uses the given container to generate the client code.
+func (t *TypescriptSdk) generateClient(ctr *Container, name, introspectionJSON, subPath string) *Directory {
+	return ctr.
+		// Add dagger codegen binary.
+		WithMountedFile(codegenBinPath, t.SDKSourceDir.File("/codegen")).
+		// Write the introspection file.
+		WithNewFile(schemaPath, ContainerWithNewFileOpts{
+			Contents: introspectionJSON,
+		}).
+		// Execute the code generator using the given introspection file.
+		WithExec([]string{
+			codegenBinPath,
+			"--lang", "typescript",
+			"--output", ModSourceDirPath,
+			"--module-name", name,
+			"--module-context-path", filepath.Join(ModSourceDirPath, subPath),
+			"--introspection-json-path", schemaPath,
+		}, ContainerWithExecOpts{
+			ExperimentalPrivilegedNesting: true,
+		}).
+		// Return the generated code directory.
+		Directory(filepath.Join(ModSourceDirPath, subPath, GenDir))
 }
 
 // DetectRuntime returns the runtime(bun or node) detected for the user's module
@@ -268,7 +287,7 @@ func (t *TypescriptSdk) Base(runtime SupportedTSRuntime) (*Container, error) {
 // If a package-lock.json, yarn.lock, or pnpm-lock.yaml is present, node will be used.
 // If a bun.lockb is present, bun will be used.
 // If none of the above is present, node will be used.
-func (t *TypescriptSdk) DetectRuntime(ctx context.Context, modSource *ModuleSource, subPath string) (SupportedTSRuntime, error) {
+func (t *TypescriptSdk) detectRuntime(ctx context.Context, modSource *ModuleSource, subPath string) (SupportedTSRuntime, error) {
 	// Try to detect runtime from package.json
 	source := modSource.ContextDirectory().Directory(subPath)