Following https://github.com/dagger/dagger/issues/7093, I found that most of the time is spent downloading dependencies, it seems I can drastically reduce the download time using `yarn` and the `--production` flag to only install production dependencies.

This is a first step toward improving TS performances, we might switch to `pnpm` later, this require some tests though because I'm hitting an issue (see https://github.com/dagger/dagger/issues/7093#issuecomment-2057038646)