@TomChv 2024-05-14
It seems theses changes in the setup introduce issues in the tests, sometime the dependencies cannot be installed properly and it's a weird typescript error.
I'm digging into it