@TomChv 2024-05-14
I can reproduce the error using `dagger call --source=. test custom --run="TestModuleTypescriptInit"`

I'm going to try to reproduce it locally, it seems something is going wrong when using `dagger init` with the new setup, I'll update this comment with my researches.