@helderco 2024-05-15
> I saw that I copy the `runtime` directory into base container, it may be possible to improve performances by ignoring part of it. I think I could apply the same logic in the `sdk` directory to ignore tests files that are useless during the execution.

You may have forgotten but putting the whole `sdk` directory was just a stop gap so we could work on more important things first. But the intent was to eventually cycle back and refactor the SDK to decouple the parts that aren't needed in modules. Python needs the same.