@gerhard 2024-05-16
I would very much like to see a version of this land in the next milestone, `v0.11.5`:
- https://github.com/dagger/dagger/milestone/47 

Even if we don't finish everything, any improvement that we can make will go a long way.