@gerhard 2024-05-21
I tried this with a simple TypeScript pipeline for a Next.js app:
```typescript
/**
 * Build automation powered by Dagger
 */
import { dag, Directory, Secret, object, func } from "@dagger.io/dagger"

@object()
class App {
  /**
   * Publish image: dagger call image --source=.:app --token=env:GH_TOKEN
   */
  @func()
  async image(source: Directory, token: Secret): Promise<string> {
    return dag
      .container()
      // https://hub.docker.com/_/node/tags?page=&page_size=&ordering=&name=lts
      .from("node:lts-iron")
      // https://github.com/krallin/tini
      .withExec(["apt", "update"])
      .withExec(["apt", "install", "--yes", "tini"])
      .withDirectory("/app", source)
      .withWorkdir("/app")
      .withExec(["npm", "install"])
      .withExec(["npm", "run", "build"])
      .withEntrypoint(["tini", "--", "npm", "run", "start"])
      .withRegistryAuth("ghcr.io", "gerhard", token)
      .publish("ghcr.io/gerhard/nextjs")
  }
}
```

I am connecting to a remote Docker instance via SSH:
```console
Client: Docker Engine - Community
 Version:           26.0.0
 OS/Arch:           darwin/amd64

Server: Docker Engine - Community
 Engine:
  Version:          26.1.3
  OS/Arch:          linux/amd64
 containerd:
  Version:          1.6.31
 runc:
  Version:          1.1.12
```

Linux 6.8 x86_64 host where Docker / Dagger is running has:
- 16 CPUs
- 32GB DDR5
- NVMe disk

These are the results for various runs:

| COMMAND                         | `v0.11.4` - BEFORE THIS CHANGE | `dev` - AFTER THIS CHANGE |
| -                               | -                              | -                         |
| 1st run - no cache              | `95s`                          | `108s`                    |
| 2nd run - cached                | `11s`                          | `9s`                      |
| 3rd run - `touch foo`           | `50s`                          | `49s`                     |
| 4th run - module comment change | `32s`                          | `19s`                     |
| 5th run - module comment change | `22s`                          | `20s`                     |

Based on the above measurements, I observed one improvement, **4th run**, when I just kept updating a comment in the module, which was `40%` quicker: `19s` vs `32s`

Before this change, that was inconsistent: **5th** run was only `10%` quicker: `20s` vs `22s`

I also observed one deterioration: **1st run - no cache** was `13%` slower: `108s` vs `95s`

Here are Dagger Cloud traces for the interesting runs above:
- `1st run - no cache` [before this change](https://dagger.cloud/dagger/traces/0547c97176228e750c767ed52763aef6) & [after this change](https://dagger.cloud/dagger/traces/db5c8ee47a3539fadc63e8b1d61e06a3)
- `4th run` [before this change](https://dagger.cloud/dagger/traces/f2756499c07b581dd654952459c55e40) & [after this change](https://dagger.cloud/dagger/traces/143b8817ea96ca6abb89fcb8135b9460)
- `5th run` [before this change](https://dagger.cloud/dagger/traces/39e3f02038b8866eb2acc9ab4d1cee66) & [after this change](https://dagger.cloud/dagger/traces/806b48cbbad8676a3c429f1ffe2486e4)

Based on the above measurements, this does not seem a net performance improvement to me. What do you think @TomChv:
1. Do my measurements look correct?
2. Was I measuring the right thing?
3. Do you still think that it's worth continuing in this direction?