@TomChv 2024-05-21
Thanks for your amazing feedbacks! It really helps me understand what's going on with a real world use case.

> Do my measurements look correct?

Yep absolutely, I could hit the same kind of behaviour and flakiness on my tests! Specially between 4 and 5.

The improvement of 40% to 10% is also something I got which confirm my work! I could hit a 2x improvement with an old version but it seams that it wasn't supporting every cases (to be more specfic: the `hasPkgJson` & `hasTsConfig`) was failing.

>Was I measuring the right thing?

Yes! Perfect, I needed the run and the traces, to see how much time take every runs and which steps is longer.

>Do you still think that it's worth continuing in this direction?

I think what I've done in this PR is a good start for further improvement, there's still some mystery like this example below where I cannot understand how is that possible.

<img width="1404" alt="Screenshot 2024-05-21 at 21 48 54" src="https://github.com/dagger/dagger/assets/61683879/70185d83-909e-4a04-95b1-6a910b82ba6a">

I'm super confused about the fact that yarn was waaay slower than npm when the `withExec` has multiple instruction, and it's even weirder because npm is only supposed to update the `package.json` (that's why `--package-lock-only` is done) but then the real [`npm install` took only 2.3s](https://dagger.cloud/dagger/traces/0547c97176228e750c767ed52763aef6#e3eb0339e5509fb3) which is super confusing because it should have install all the package.

I think I'm missing something about yarn, which keep us from using the cache. I'm thinking that at some point, we could hit higher performance if we were running the TS SDK in runtime language but that seams quite complex to do.
The overall installation of dependencies takes time, the cache isn't hitting every time (which doesn't make sense since nothing should have before adding user's source in `CodegenBase`.

I think it's still worth merging the PR, because it makes the module much clearer but there's still a long way to go before getting a super fast TS SDK