@TomChv 2024-05-21
I'm also having a concern about the CI and the flakiness this PR introduce, on multiple runs, I could see different tests passing & failing, with always this kind of error:

```
couldn't find a package.json file in "/usr/local/share/.cache/yarn/v6/npm-@dagger-io-dagger-0.0.0-1cdff291-b199-40fc-9ef6-3ff1ac6d6d1d-1716293036796/node_modules/@dagger.io/dagger" 
```

Meaning that there's an issue with yarn cache and I cannot understand why is that so, this is something that isn't happening with npm at all.