@TomChv 2024-05-21
By doing so, I could successfully get a green success on initialization, which is a good news

<img width="1230" alt="Screenshot 2024-05-21 at 23 46 44" src="https://github.com/dagger/dagger/assets/61683879/6cf16864-94c4-4a2a-9345-ad8e359b5065">