@TomChv 2024-05-21
@gerhard It seems I could fix the CI, could you do one more benchmark with your example just to see if there's any changes? I think it might get slower, I'll need to find why the yarn cache fails with collision.
A really simple check would be to add a specific cache-key for each module but I'm not sure we want to do that.